# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spicy Noodle Management API",
    "version": "1.0.0",
    "description": "REST API cho hệ thống quản lý nhà hàng mỳ cay. Docs được generate từ annotations trong code."
  },
  "servers": [
    {
      "url": "https://seoul-spicy-production.up.railway.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000/api",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Đăng nhập",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {"type": "string", "example": "0123456789"},
                  "password": {"type": "string", "example": "Pass@123"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}